+++ IR/transmitter/cpp/hw_if/pico_hardware_interface.h
@@ -1,5 +1,6 @@
 #pragma once
-
+#include "ihardware_interface.h"
+#include "hardware_interface.h"
 
 class PicoHardwareInterface : public IHardwareInterface
 {
@@ -15,7 +16,7 @@ class PicoHardwareInterface : public IHardwareInterface
     virtual int get_pins();
     virtual int wait_pins_change(sPinsChangeData* pinsChangeData, int mask, uint64_t timeout);
     virtual void gpio_put(int pin_number, bool value);
-    virtual int gpio_put_at_us(int pin_number, bool value, int time_us);
+    virtual uint64_t gpio_put_at_us(int pin_number, bool value, int time_us);
     virtual void set_led(bool value);
     virtual void sleep_us(int useconds);
     virtual int tcp_request(const char* server, unsigned int port, const char* request, std::string& response);
